Refactor `utils`
