+++ ascent/predict.py
@@ -153,8 +153,8 @@ class AscentPredictor(AscentTrainer):
 
         Currently, this method only supports inference for nnUNet models.
 
-        This method is wrapped in optional @task_wrapper decorator which applies extra utilities
-        before and after the call.
+        This method is wrapped in optional @task_wrapper decorator, that controls the behavior during
+        failure. Useful for multiruns, saving info about the crash, etc.
 
         Args:
             cfg (DictConfig): Configuration composed by Hydra.
@@ -165,6 +165,10 @@ class AscentPredictor(AscentTrainer):
         Raises:
             ValueError: If the checkpoint path is not provided.
         """
+        # apply extra utilities
+        # (e.g. ask for tags if none are provided in cfg, print cfg tree, etc.)
+        utils.extras(cfg)
+
         if not cfg.ckpt_path:
             raise ValueError("ckpt_path must not be empty!")
