+++ ascent/utils/__init__.py
@@ -1,12 +1,5 @@
+from ascent.utils.instantiators import instantiate_callbacks, instantiate_loggers
+from ascent.utils.logging_utils import log_hyperparameters
 from ascent.utils.pylogger import get_pylogger
 from ascent.utils.rich_utils import enforce_tags, print_config_tree
-from ascent.utils.utils import (
-    close_loggers,
-    extras,
-    get_metric_value,
-    instantiate_callbacks,
-    instantiate_loggers,
-    log_hyperparameters,
-    save_file,
-    task_wrapper,
-)
+from ascent.utils.utils import extras, get_metric_value, task_wrapper
